requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.3
defusedxml==0.7.1
Django==5.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
flake8==7.1.1
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
py==1.11.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pyparsing==3.2.1
pytest==8.3.3
pytest-django==4.9.0
python3-openid==3.2.0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.2
toml==0.10.2
urllib3==2.4.0
